Added files via upload
These are the index and style sheet for the website. This is a one page website, and the images used are located in a local machine, so will not show up in the code. Currently in the process of getting the images, pdfs, and website all online so that it can be viewed by people.